refactor: move file tree models to separate module
