+++ backend/server/src/config.rs
@@ -1,25 +1,7 @@
 use anyhow::Result;
 use redis::{Client, aio::ConnectionManager};
 use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
 use tokio::fs;
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct FileInfo {
-    pub file_path: String,
-    pub upload_timestamp: u64,
-    pub file_size: u64,
-}
-
-pub type BucketFiles = Vec<FileInfo>;
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum NodeType {
-    File(FileInfo),
-    Node(TreeNode),
-}
-
-pub type TreeNode = HashMap<String, NodeType>;
 
 #[derive(Debug, Clone, Deserialize)]
 pub struct Settings {
